Teapot: Sonarcube fix
-remove duplicated code
