+++ Teapot/src/ShapeGenerator/Cube.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Shapes.h"
-#include "Utils.h"
 
 namespace Shapes
 {
@@ -22,6 +21,5 @@ namespace Shapes
 		const glm::vec3 m_color;
 
 		std::vector<glm::vec3> m_unitVertices;
-		Utils m_utils;
 	};
 }
